Harden Node file scheduled work
