+++ packages/runtime/src/platform/node/host/file-execution-host.test.ts
@@ -19,6 +19,7 @@ import {
 } from "../../../testing/test-fixtures";
 import { collect } from "../../../thread/handle/test-support";
 import { createNodeFileExecutionHost } from "./file-execution-host";
+import { drainScheduledNodeWork } from "./scheduled-work-drainer";
 import {
   ackScheduledNodeRun,
   ackScheduledNodeThreadPrompt,
@@ -26,7 +27,7 @@ import {
   appendScheduledNodeThreadPrompt,
   listScheduledNodeRuns,
   listScheduledNodeThreadPrompts,
-} from "./scheduled-work-queue";
+} from "./scheduled-work-store";
 
 const malformedScheduledWorkPattern =
   /Invalid Node scheduled work file .*invalid JSON/;
@@ -59,13 +60,15 @@ describe("createNodeFileExecutionHost", () => {
       ]);
 
       const secondHost = createNodeFileExecutionHost({ directory });
-      const agent = new Agent({
-        host: secondHost,
-        model: createCallbackModel(() =>
-          Promise.resolve([assistantMessage("RESUMED")])
-        ),
-        namespace: "local-owner",
-      });
+      const createAgent = () =>
+        new Agent({
+          host: secondHost,
+          model: createCallbackModel(() =>
+            Promise.resolve([assistantMessage("RESUMED")])
+          ),
+          namespace: "local-owner",
+        });
+      const agent = createAgent();
 
       expect(agent.supportsResume).toBe(true);
 
@@ -97,15 +100,28 @@ describe("createNodeFileExecutionHost", () => {
       const duplicate = await agent.resume(dispatched.runId);
       expect(duplicate).toBeNull();
 
+      const drainResult = await drainScheduledNodeWork({
+        agentForRun: () => createAgent(),
+        directory,
+      });
+      expect(drainResult.ackedThreadPrompts).toEqual([
+        expect.objectContaining({
+          notificationId: dispatched.notificationId,
+          runId: dispatched.runId,
+          threadKey: "thread:local",
+        }),
+      ]);
+      expect(drainResult.events).toEqual([]);
+
       const dataDirectory = await currentDataDirectory(directory);
       expect(await readdir(join(dataDirectory, "threads"))).not.toHaveLength(0);
       expect(await readdir(join(dataDirectory, "runs"))).not.toHaveLength(0);
       expect(
         await readdir(join(dataDirectory, "notifications"))
       ).not.toHaveLength(0);
-      await expect(
-        readdir(join(directory, "scheduled-work", "thread-prompt"))
-      ).resolves.toHaveLength(1);
+      await expect(listScheduledNodeThreadPrompts(directory)).resolves.toEqual(
+        []
+      );
     } finally {
       await rm(directory, { force: true, recursive: true });
     }
@@ -114,6 +130,9 @@ describe("createNodeFileExecutionHost", () => {
   it("dedupes and acks local scheduled work files", async () => {
     const directory = await tempDir();
     try {
+      await appendScheduledNodeRun(directory, "run:delayed", {
+        runAfterMs: 1000,
+      });
       await appendScheduledNodeRun(directory, "run:1");
       await appendScheduledNodeRun(directory, "run:1");
       await appendScheduledNodeThreadPrompt(directory, {
@@ -132,6 +151,9 @@ describe("createNodeFileExecutionHost", () => {
       await expect(listScheduledNodeRuns(directory)).resolves.toEqual([
         "run:1",
       ]);
+      await expect(
+        listScheduledNodeRuns(directory, { nowMs: Date.now() + 2000 })
+      ).resolves.toEqual(["run:1", "run:delayed"]);
       const prompts = await listScheduledNodeThreadPrompts(directory);
       expect(prompts).toHaveLength(1);
       expect(prompts[0]).toEqual({
@@ -148,6 +170,60 @@ describe("createNodeFileExecutionHost", () => {
       await expect(listScheduledNodeThreadPrompts(directory)).resolves.toEqual(
         []
       );
+    } finally {
+      await rm(directory, { force: true, recursive: true });
+    }
+  });
+
+  it("keeps scheduled work pending when the run is still leased", async () => {
+    const directory = await tempDir();
+    try {
+      const host = createNodeFileExecutionHost({ directory });
+      const agent = new Agent({
+        host,
+        model: createCallbackModel(() =>
+          Promise.resolve([assistantMessage("SHOULD NOT RUN")])
+        ),
+        namespace: "local-owner",
+      });
+      await host.store.runs.create({
+        checkpointVersion: 0,
+        kind: "notification",
+        lease: {
+          attempt: 1,
+          leaseId: "active-lease",
+          leaseUntilMs: Date.now() + 60_000,
+        },
+        ownerNamespace: agentNamespace("local-owner"),
+        rootRunId: "run:leased",
+        runId: "run:leased",
+        status: "leased",
+        threadKey: "thread:leased",
+      });
+      const prompt = {
+        idempotencyKey: "notify:leased",
+        notificationId: "notification:leased",
+        runId: "run:leased",
+        threadKey: "thread:leased",
+      };
+      await appendScheduledNodeRun(directory, "run:leased");
+      await appendScheduledNodeThreadPrompt(directory, prompt);
+
+      const drainResult = await drainScheduledNodeWork({
+        agentForRun: () => agent,
+        directory,
+      });
+
+      expect(drainResult.ackedRuns).toEqual([]);
+      expect(drainResult.ackedThreadPrompts).toEqual([]);
+      expect(drainResult.skippedRuns).toEqual(["run:leased"]);
+      expect(drainResult.skippedThreadPrompts).toEqual([prompt]);
+      await expect(listScheduledNodeRuns(directory)).resolves.toEqual([
+        "run:leased",
+      ]);
+      await expect(listScheduledNodeThreadPrompts(directory)).resolves.toEqual([
+        prompt,
+      ]);
     } finally {
       await rm(directory, { force: true, recursive: true });
     }
