+++ packages/runtime/src/platform/node/storage/file-execution-store/types.ts
@@ -1,0 +1,3 @@
+export type DataDirectoryResolver = () => Promise<string>;
+
+export type FileExecutionLock = <T>(fn: () => Promise<T>) => Promise<T>;
