+++ packages/runtime/src/platform/node/storage/file-execution-store/utils.ts
@@ -1,0 +1,11 @@
+export function encodeKey(key: string): string {
+  return Buffer.from(key).toString("base64url");
+}
+
+export function isNodeError(error: unknown): error is NodeJS.ErrnoException {
+  return error instanceof Error && "code" in error;
+}
+
+export function isRecord(value: unknown): value is Record<string, unknown> {
+  return value !== null && typeof value === "object" && !Array.isArray(value);
+}
