+++ scripts/verify-release-artifacts-node.test.mjs
@@ -45,7 +45,11 @@ describe("verifyReleaseArtifacts node declaration checks", () => {
       "packages/runtime/dist/platform/node/index.d.ts: missing explicit node runtime export NodeFileExecutionHostOptions",
       "packages/runtime/dist/platform/node/index.d.ts: missing explicit node runtime export NodeFileThreadHostOptions",
       "packages/runtime/dist/platform/node/index.d.ts: missing explicit node runtime export NodeScheduledThreadPrompt",
+      "packages/runtime/dist/platform/node/index.d.ts: missing explicit node runtime export NodeScheduledWorkAppendOptions",
+      "packages/runtime/dist/platform/node/index.d.ts: missing explicit node runtime export NodeScheduledWorkDrainOptions",
+      "packages/runtime/dist/platform/node/index.d.ts: missing explicit node runtime export NodeScheduledWorkDrainResult",
       "packages/runtime/dist/platform/node/index.d.ts: missing explicit node runtime export NodeScheduledWorkListOptions",
+      "packages/runtime/dist/platform/node/index.d.ts: missing explicit node runtime export NodeScheduledWorkRunContext",
       "packages/runtime/dist/platform/node/index.d.ts: missing explicit node runtime export ackScheduledNodeRun",
       "packages/runtime/dist/platform/node/index.d.ts: missing explicit node runtime export ackScheduledNodeThreadPrompt",
       "packages/runtime/dist/platform/node/index.d.ts: missing explicit node runtime export appendScheduledNodeRun",
@@ -54,6 +58,7 @@ describe("verifyReleaseArtifacts node declaration checks", () => {
       "packages/runtime/dist/platform/node/index.d.ts: missing explicit node runtime export createNodeFileExecutionHost",
       "packages/runtime/dist/platform/node/index.d.ts: missing explicit node runtime export createNodeFileScheduler",
       "packages/runtime/dist/platform/node/index.d.ts: missing explicit node runtime export createNodeFileThreadHost",
+      "packages/runtime/dist/platform/node/index.d.ts: missing explicit node runtime export drainScheduledNodeWork",
       "packages/runtime/dist/platform/node/index.d.ts: missing explicit node runtime export listScheduledNodeRuns",
       "packages/runtime/dist/platform/node/index.d.ts: missing explicit node runtime export listScheduledNodeThreadPrompts",
     ]);
